more changes to everything
