+++ enzy_htp/structure/structure_constraint/__init__.py
@@ -26,7 +26,7 @@ from .create_constraint import (
     create_angle_constraint,
     create_dihedral_constraint,
     merge_cartesian_freeze,
-    freeze_hydrogen_bonds 
+    create_hydrogen_bond_freeze 
     )
 
 from .xml_io import (
