+++ test/structure/structure_region/test_capping.py
@@ -38,7 +38,6 @@ def test_capping_with_residue_terminals(helpers):
     assert len(test_region.atoms) == len(unique_names)
         
 
-
 def test_all_cap_types_supported():
     """Checking that all the cap types that are supposed to work do not throw errors."""
     test_stru = sp.get_structure(f"{DATA_DIR}tri_alanine.pdb")
@@ -206,3 +205,41 @@ def test_correct_num_caps():
 
     assert len(caps) == 2 
 
+def test_capping_with_residue_terminals(helpers):
+    """as name. use an example region from selection.
+    Just make sure not execptions is raised for now.
+    TODO complete this test"""
+    test_stru = sp.get_structure(f"{DATA_DIR}KE_07_R7_2_S.pdb")
+    sele = stru_sele.select_stru(
+        test_stru, "resi 1+2+3+7+8")
+    test_region = stru_regi.StructureRegion(atoms=sele.atoms)
+    capping.capping_with_residue_terminals(test_region, nterm_cap='CH3', cterm_cap='CH3')
+    # test
+    atoms = test_region.atoms
+    test_file = f"{WORK_DIR}test_capping_1.xyz"
+    answer_file = f"{DATA_DIR}answer_capping_1.xyz"
+    lines = [
+        str(len(atoms)),
+        "",]
+    for at in atoms:
+        lines.append(f"{at.element} {at.coord[0]} {at.coord[1]} {at.coord[2]}")
+    fs.write_lines(test_file, lines)
+    assert helpers.equiv_files(test_file, answer_file, consider_order=False)
+
+    fs.safe_rm(test_file)
+
+def test_residue_cap_deepcopy():
+    """test the behavior of deepcopy for residue cap"""
+    test_stru = sp.get_structure(f"{DATA_DIR}KE_07_R7_2_S.pdb")
+    test_res = test_stru.residues[0]
+    test_res_2 = test_stru.residues[1]
+    test_ch3 = capping.get_ch3_cap("cterm", test_res, test_res.atoms[0], test_res_2.atoms[0])
+
+    ch3_copy = deepcopy(test_ch3)
+
+    test_region = stru_regi.StructureRegion(atoms=test_res.atoms) #return_copy
+    capping.capping_with_residue_terminals(test_region, return_copy=True)
+    assert ch3_copy.link_residue is not test_res
+    assert ch3_copy.link_residue.parent is None
+    assert ch3_copy.plug_atom in ch3_copy.atoms
+    assert ch3_copy.link_atom in ch3_copy.link_residue.atoms
